+++ app/src/parser_txdef.h
@@ -158,6 +158,7 @@ typedef struct {
         candid_ListNeurons_t candid_listNeurons;
         SendRequest SendRequest;
         ListNeurons ListNeurons;
+        sns_ManageNeuron_t sns_manageNeuron;
     } data;
 } call_t;
